requirements.txt
pyserial>=3.5
websockets>=12.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
